backend/requirements.txt
# =============================================================
# requirements.txt — Smart Agriculture Backend Dependencies
# Install with: pip install -r requirements.txt
# Compatible with Python 3.13+
# =============================================================

# ── Web Framework ─────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1      # ASGI server to run FastAPI

# ── MQTT Client ───────────────────────────────────────────────
paho-mqtt==2.0.0               # MQTT subscriber for FastAPI

# ── Data Validation ───────────────────────────────────────────
pydantic==2.10.3               # Data models & validation (built into FastAPI)
pydantic-settings==2.6.1       # .env file loading into settings

# ── Database ──────────────────────────────────────────────────
motor==3.6.0                   # Async MongoDB driver (for Phase 3)
pymongo==4.9.2                 # Must be <4.10 — required by motor 3.6.0

# ── Utilities ─────────────────────────────────────────────────
python-dotenv==1.0.1           # Load .env file
python-dateutil==2.9.0         # Date/time utilities

# ── Development ───────────────────────────────────────────────
httpx==0.27.0                  # For testing API endpoints
